+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   * Bug fix when local config file was missing in `DefaultBootstrap::launchFirst`.
   * Changed how `--delegate` argument in `Bootstrap` works.
   * Fixed bug in `deleteOldFiles()` where it would try to compare files even if it belongs to different OS.
+  * `Configuration.getTimestamp()` now returns `null` if not present in the XML instead of time of object creation.
 * **1.3.3**
   * Properly escape special chars in output XML. Allow control chars in properties.
   * Renamed `UpdateHandler::connect` to `UpdateHandler::openDownloadStream`